🧰: move `SelectionElement` and `Selection` into `top-bar.cp.js`
This is meant to improve the separation of concerns between the IDE world and the top-bar. Both have functionality related to halos. However, the Selection behavior belongs solely to the top-bar, and this should be reflected in the code organisation.
